+++ src/components/LoadingOverlay.tsx
@@ -1,7 +1,7 @@
 // src/components/LoadingOverlay.tsx
 import React, { useEffect, useRef } from 'react';
 import { View, Text, StyleSheet, Animated } from 'react-native';
-import { colors } from '@theme/colors';
+import { colors } from '../theme/colors';
 
 type Props = {
   title?: string;
@@ -25,7 +25,7 @@ export const LoadingOverlay: React.FC<Props> = ({ title = 'ClassiQ', subtitle })
   return (
     <View style={styles.container} pointerEvents="none">
       <Animated.View style={{ transform: [{ scale }] }}>
-        <Text style={styles.title}>ClassiQ</Text>
+        <Text style={styles.title}>🎼 ClassiQ</Text>
         {subtitle ? <Text style={styles.subtitle}>{subtitle}</Text> : null}
       </Animated.View>
     </View>
@@ -41,14 +41,16 @@ const styles = StyleSheet.create({
     zIndex: 9999,
   },
   title: {
-    fontSize: 54,
-    fontWeight: '800',
+    fontFamily: 'Lora_700Bold',
+    fontSize: 30,
+    fontWeight: 'bold',
+    marginBottom: 60,
     letterSpacing: 0.5,
     color: colors.textOnSplash,
   },
   subtitle: {
-    marginTop: 8,
-    fontSize: 14,
+    marginTop: 15,
+    fontSize: 20,
     color: colors.textOnSplash,
     opacity: 0.8,
     textAlign: 'center',